Response to SQL Diagnostic
Comfort: 1/5
Clarity: 4/5

I had no background in SQL and find it difficult because the repo for
SQL does not include very many explicit instructions on how to reproduce
the things we did in class. I had no issues at the time but simply don't
remember how to do a lot of it.