SQL Diagnostic
Comfort: 3
Clarity: 3

I felt comfortable following along in the lesson but needed to review
materials after not using it for a few days. If I had more time I would
have liked to have spent more time playing around and testing it. I
am not familiar with the syntax but am ok with the concepts.