SQL diagnostic - Tony Leo
Struggled with this one. Still not sure why I had to make a primary key
as I was pretty sure this was done for us. So, that blew up on my tests.

Did the best I could but wasn't sure if I was to be using postgresql
or just plain ole SQL.

Pretty sure most of these are wrong.